storage: Move user change post() call after user dir creation.
[storage] Move user change post() call after user dir creation. JB#55593

In order to ensure that the settings file can be created to an existing
dir move the post() call after the dir has been created. Both connmand
and vpnd use this functionality and both have separate post() functions,
although vpnd does not set such function as of now.

In order to make the saving and permission change errors more visible
change DBG() to connman_error() in storage_load() and storage_save() to
follow the approach used already with storage_delete().