storage: Do not process ENOENT in path validation when loading.
[storage] Do not process ENOENT in path validation when loading. JB#52090

Add a boolean configuration option for path validation to allow service
loading differentiate between a real error and a non-existent path. WiFi
services that are available in scan are attempted to be loaded and the
warning should not be put into logs in these cases. Instead, silently
return NULL in storage_load() as indication that loading is not
possible.