openvpn: Handle script context in notify and check signature
[openvpn] Handle script context in notify and check signature. JB#55889

Add handling of the script context in OpenVPN notify function. The
script for OpenVPN sends also the context provided by the binary and it
is to be handled prior to the environment values as a distinct
indication whether initialization or restart is taking place. In case of
any other than "init" propagate ECONNRESET error through the callback to
vpn-provider.c and forcefully stop the task and make VPN disconnected.
This is due to the fact that restarts are only partially supported yet
and the routes are not changed if VPN provider provides new ones with
the PUSH reply after internal reset of OpenVPN is completed.

Having uncontrolled restarts had issues with TCP and a misconfigured
connection using UDP may also reset itself with SIGUSR1 and repeditely
call the up script which would lead up piling of new routes. UDP does
not have such option to prevent connection retries so in such case
processing context works. User is also capable of preventing this by
remapping the SIGUSR1 to something other but this approach is a fallback
in case the SIGUSR1 is still used with OpenVPN and the approach here
applies to both TCP and UDP, since --connect-retry-max is not affecting
the case seen here with a potentially misconfigured VPN.

This was seen to happen with some UDP VPNs that use and require
--comp-lzo, which was omitted from options as an error (and fixed in
this patch-set), and then the binary reports:
  write to TUN/TAP : Invalid argument (code=22)
and as a result, after the ping-restart timeout has been reached,
OpenVPN does an internal restart and informs it via up script.

Add signature checking into notify and report unknown state if fails.
The unknown state is ignored by plugins/vpn.c and can be utilized as a
notify for invalid message. Even though the messages to notify should be
arriving only from the script it is imperative to ensure that invalid
content is not being sent and processed.