[ofono] Support 3GPP TS 24.301 dual network mode. Fixes JB#56481
The bailout timer fix 6bc974c039e1a9db9b38584bfbc4df3746b04f1d
introduced a late connection of the network if either of the IP family
configurations was not set yet. This did not account for the situation
properly when either of these IP family configurations are completely
missing according to 3GPP TS 24.301:
 "if the UE requests for PDN type IPv4v6, but the subscription is limited
  to IPv4 only or IPv6 only for the requested APN, the network shall
  override the PDN type requested by the UE to be limited to a single
  address PDN type (IPv4 or IPv6)."

This change affects only dual mode and when the missing IP configuration
is waited for the period set in bailout timer fix the missing one is set
to OFF to avoid disconnection of mobile data. This makes the plugin to
work in dual mode mobile data where operator does not deliver both IP
configurations.

This fixes the issue of looping a cellular connection from
connecting->connected->error->idle->connecting->connected when either of
the IP family types are not sent by the operator or detected by ofono.