[wifi] Implement a heurestic for determening a weak WiFi. JB#56225 OMP#JOLLA-552
This adds a heurestic for determining whether the WiFi network is weak
or not based on the amount of services there is. The weak signal
checking is now increasing as the WiFi services accumulate. This has an
effect in to how the services are added and thus, in crowded areas will
reduce the amount of really weak WiFis. The signal strength comparison
starts from -85 dBm up to the background scanning limit -65dBm. This
change makes the value to increase by 5 dBm by every 20 WiFi networks
that are available. Thus, background scanning limit is reached at 80
networks.

Main point here is to keep only the strongest WiFi networks in the list
as the really weak ones might not even be connectable. In areas where
there are a lot of networks, lets say over 60 it will eventually be
difficult to even get connected to a -85 dBM signal network let alone
keep it connected.

This change should reduce the load on UI layers in crowded areas in
terms of WiFi networks.