dnsproxy: protocol_offset: remove error return case and return size_t
There is never any socket created that is not UDP or TCP, so this error
should never occur unless some memory corruption or similar occured.
Therefore `abort()` on unexpected values instead.

This simplifies the calling code since the error case no longer needs to
be covered (which can also be dangerous when forgotten, because then
negative error codes are used in calculations).