service: Check if hidden service has a pending request on agent
Hidden (WiFi) services do not save the pending request to the service
struct and this utilizes agent-connman.c to check if there is a agent
request already set for the service. In such case this reports in
progress error back to the caller as with visible service.

Otherwise there would be a possibility to request connect for at least
twice for a hidden WiFi network. And when agent is in use this would
result in two requests to be sent, first one is the current active one
and the second would be waiting in agent queue, which in case of
time-out would be then be sent to the agent as a duplicate request.