timezone: Change regdom along timezone, use localtime config
This is an optional feature and when enabled with RegdomFollowsTimezone
config option the regdom is changed when the timezone changes. This
feature is useful in cases when the devices without a cellular
functionality or a SIM card to set the regdom for WiFi via regdom
changes indicated by Ofono plugin are remaining the default setting (US).
As a result some region specific channels cannot be utilized with WiFi.

The ISO3166 country code is read from /usr/share/zoneinfo/zone1970.tab
using the set timezone. This is done at startup and when there is an
inotify event for the localtime. The first ISO3166 country code set on
the line is used.

Timezone.c now uses localtime configured in main.conf. It may be a
symlink as external service can manage localtime so IN_CREATE event
is added for inotify watch.