dhcp: Set proxy properly when applying DHCP lease
When DHCP Option 252 (PAC file URL) is included,
the service should be updated. Doing so sets the proxy
method to 'auto', updates the ipconfig object, and
notifies PACrunner (if included) that the proxy setting
has changed.

Currently, only the ipconfig object is updated. This patch
calls the proper function to perform these actions.