wispr: Fix online check when using WPAD/PAC
A PAC file can return multiple proxies in response to a query.
They are separated by a semicolon and, possibly, a space.

'PROXY proxy1.example.com:3128; PROXY proxy2.example.com; DIRECT'

Currently, this is not handled, and the inclusion of more than
one proxy causes the online check to fail. This fix parses the
first proxy in the list and uses it.