clock: fix time update transition auto->manual
When the TimeUpdates property is changed via D-Bus from auto to manual,
the change is written out to the settings state file. However, it does
no modify the current service. This causes the service to continue to
synchronize time even after TimeUpdates being set to manual. This patch
updates the service whenever any change is made to the TimeUpdates setting.