rtln: Extend notify and better support IPv6 routes set via RA
[rtnl] Extend notify and better support IPv6 routes set via RA. JB#58105

Extend the notify to support IPv6 better when gateways are added and
removed. These use the destination as well as metric value for more
details about the routes. Metric for IPv6 is retrieved from
RTA_PRIORITY as it is used by kernel, see, e.g.,
https://github.com/torvalds/linux/blob/master/net/ipv6/route.c

Add toggle to pass also the routes set by RA (protocol type RTPROT_RA)
to the functions that do request it. For this an additional field to the
notify and a function to enable RTPROT_RA support was added. This could
be later re-designed to support the different user-specific protocol
types defined in linux/rtnetnlink.h but as of now includes only
RA-types.

This change adds two callbacks to the struct connman_rtnl while keeping
rest of the content intact. Each callback gets also the rtm_protocol
info so the responsibility of processing decision is on the component
implementing the callback. After the user-specific protocol types can be
added this proves to be quite useful in filtering incoming IPv6 netlink
messages.

Also support setting address family for the rtnl request. Keep AF_INET
for the internal rtnl.c functions but support also AF_INET6 and
AF_UNSPEC. The value is set only via internal/plugin calls when needed.

In some cases it is also useful to have route request available for the
plugins if there is a plugin monitoring a more complex network
configuration where, for example, additional not-managed interfaces
(blacklisted) are getting a default route assigned by kernel via route
advertisements and should ensure that the route is not used by default
when ConnMan assigned route is to be preferred.