blacklist monitor: Plugin to prevent IPv6 default routes on blacklisted ifs
[blacklist monitor] Plugin to prevent IPv6 default routes on blacklisted ifs. JB#58105

Introduce a new plugin to prevent non-monitored and managed interfaces
set in blacklist configuration from enabling any IPv6 default routes
when ConnMan has a service connected. The routes may be added by RA for
IPv6 through kernel, or an external service can add these for the
blacklisted interfaces, like DHCP, which is not supported yet. The same
RTPROT_BOOT and RTPROT_KERNEL messages types in addition to RTPROT_RA
are supported. Plugin requests rtnl.c to pass RTPROT_RA type messages to
be processed.

The basic mechanism is by monitor via new rtnl feature the exact IPv6
default routes that are added and removed for any blacklisted interface.
These are log them into a list that is used for excess default route
removal when there is a change on the default service or its IPv6
configuration. The list contains only one default route per index and
metric pair, gateway is changeable. If there are changes on default
service or its IPv6 configuration the list of the default routes is
cleared. If there is a default service already connected new default
routes from these blacklisted interfaces are prevented unless the
connected default service index matches the blacklisted interface. This
allows to keep the default routes of an service that is mainly managed
via plugin.

Furthermore, the blacklist monitor requests IPv6 routes when the default
service changes. If the default service change and previous default
service was using a blacklisted interface it would have a potentially
conflicting IPv6 route set before the new default service's route. By
requesting the routes, and the index has changed as default service
changes this plugin can make it sure that the services using a
blacklisted, i.e., non-managed interface, will not conflict with the
current service connection.

As a result in the case when, e.g., cellular that has support for both
IPv4 and IPv6 remains active when a WLAN is connected. ConnMan does not
manage IPv6 and relies on kernel so in kernel routing table there can be
a route left on cellular if the interface is blacklisted, and hence the
interface name usually has lower ASCII index than a 'w' the cellular
interface (e.g., rmnet_data) remains on top of the wlan interface.
Thus, traffic would go on IPv4 over WLAN and IPv6 would use mobile
data. Worse case is when WLAN does not have IPv6 at all but only IPv4
exists. That will cause some confusion on user.

In such case the existing route information is requested from kernel
and the IPv6 routes that are on blacklisted interfaces are removed from
routing table. If the ipconfig is set as manual or fixed the routes are
kept in a restore list, which is utilized when the interface is used by
a default service. This is imperative because when a network with either
of those ipconfig methods have their routes removed when, e.g., WLAN is
connected they are not restored from anywhere after WLAN is dropped out.
This change ensures that they are added back once the service has become
the default service from which the route was removed earlier. Therefore,
resuming connection on SLAAC or DHCPv6 enabled connection remains
unaffected.