[wispr] Treat authentication_done as error in browser reply. JB#59298
The value of authentication_done defines the state of the D-Bus
communication between connmand and the agent. When there is an error the
cleanup should be always done. The reason for earlier use of early
return when authentication_done  == false may have been a
misunderstanding. Earlier the freeing would have been skipped for a
real error as well.

The reply to request_browser_reply comes from the browser request via
agent-connman.c:request_browser_reply() called as the callback for the
agent.c:agent_receive_message() callback set for the actual agent
message that is delivered by agent.c:agent_send_next_request(). The
errors that are listed in agent-api.txt just declare only one error
net.connman.Agent.Error.Canceled, but according to the code timeout
errors and operation canceled in the case of missing reply are handled.
This should not be relying on some arbitrary messages to determine
authentication state.

Thus it is better to follow the style upstream uses. If this needs to
define whether the authentication is a success or not it should be
implemented as a complete feature. Now the variable has an odd name but
changing at this point may not provide any added value.