vpn: Add association state before association state
[vpn] Add association state before connect state. JB#59447

This changes the state machine by adding the VPN_STATE_ASSOCIATION to be
entered right after connect() callback is called. This is needed in
order to properly react with the user input dialog waiting on VPNs.
CONNECT state is now set when the dialog is closed to indicate that user
input is given and now the VPN really connects.

When VPN notify() allback is called the connect state is enforced if the
return value indicates so and the internal state is different. This is
to accommodate the changes required and to operate as a fallback that
the states of provider and driver are kept in sync.

Handle also association state in vpn_notify() where the state reported by
plugin can only be transitioned back to association from connect state.
The initial state for the plugins is association so in that case there
is no state change done. If this state is notified back in any other
state it is logged as a warning with the plugin that caused the state
transition attempt.