[connman] Drop error use from IPv4 status URL conf check. JB#59564
GError was not cleared after status URL checks are done when checking
the configuration file. This caused an obvious error when the
configuration file did not have the status URLs defined:

(connmand:1070): GLib-WARNING **: 12:19:38.963: GError set over the top
of a previous GError or uninitialized memory. This indicates a bug in
someone's code. You must ensure an error is NULL before it's set. The
overwriting error message was: Key file does not have key ?Enable6to4?
in group ?General?

As the error was already ignored for IPv6 we can drop it also for IPv4
case hence the default is returned in connman_setting_get_string() when
the ipv{4,6}_status_url is not set (NULL).