[ofono] Increase delayed connect wait to 5 seconds. JB#58538 JB#59714
It seems that in some cases with VoLTE the too short wait for IPv4
connection information from the operator may be the cause for not
having IPv4 connectivity at all. This fix was done previously to address
certain error situations with some operators and the limit was set to 2
seconds, it might be that with VoLTE we need longer wait for the IPv4
information in dual mode.