vpnc: Ensure that vpnc_connect_done() is called at shutdown
[vpnc] Ensure that vpnc_connect_done() is called at shutdown. JB#55105

The vpnc_connect_done() is called only once, at success or at failure.
Adding this call makes sure that it will be called when the server
returns invalid response, configuration is incompatible with the server
or vpnc cannot be run on the system (e.g., vpnc is ran as regular user
that has no permission to create raw sockets). By doing so the D-Bus
LimitsExceeded can be avoided as every pending request, even in case of
failures, will get a response.