service: Handle provider EALREADY replies if connecting a VPN
[service] Handle provider EALREADY replies if connecting a VPN. JB#55105

The provider.c now returns EALREADY when trying to connect a VPN that is
already being set to connect. In case of VPN autoconnection this should
be dealt similarly to EINPROGRESS as it indicates we could have a VPN
left to connect as the result of connection is not yet known.

Also cleanup the service/provider connect reply code handling. By
reacting to EALREADY the connect_timeout() is being set only once for a
connection attempt.