vpn: Add VPN agent use callback for plugins
[vpn] Add VPN agent use callback for plugins. JB#59447

Add callback that can be used by the VPN plugins to tell the vpn_driver
whether it uses VPN agent or not. Default to using VPN agent if the
function is not defined.

This is done to accommodate the state transition in vpn-provider when
the VPN does not utilize VPN agent.