vpn-provider: Transition to CONNECT state with agentless VPNs
[vpn-provider] Transition to CONNECT state with agentless VPNs. JB#59447

Set to transition to CONNECT state immediately after ASSOCIATION state
after initializing connection procedure with a VPN if it does not use
VPN agent. This is done to accommodate the full state machine
transitions as the VPN agent, when success, will do the transition but
when VPN agent is not used the transition would be required to be done
by the plugin.