service: Fix an additional case of autoconnect breakage under iwd
Error flow:
- __connman_service_create_from_network() creates network, but it is not
  favorite since it is not yet saved
- trigger_autoconnect is skipped since !favorite

*** User calls Connect() ***
- iwd plugin gets connect() call
- iwd plugin finds out about new known network and copies the
  autoconnect state of the iwd_network to the iwd_known_network, which
  is false since it was zero initialized and nobody called
  connman_network_set_autoconnect on it.
- iwd plugin tells iwd to set AutoConnect to false
- service_indicate_state() calls __connman_service_set_favorite(service,
  true)
- nobody tells the iwd plugin about the change in favorite state

This patch calls trigger_autoconnect right at the end there, which will
then propagate the outcome of the favorite state to the iwd plugin. I
think this patch is the right design for the current architecture of the
autoconnect state management, but I think the autoconnect design
probably needs some changes outside the scope of this patch since it is
too easy to make these mistakes.