[clat] Fix error cases, add leeway + tolerance for resolv errors. JB#52923
The errors requiring a restart and shutdown are changed to always go
through task stopping to avoid jumping over a state improperly. This
ensures that the tayga process is waited for to quit before moving to
next state.

Added tolerance for resolv errors similarly to timeouts. The retry is
now set to be conducted every 2 seconds in both cases for 6 attempts.
This is to mitigate the potential network setup delay and to allow
tolerance for resolving errors while running and to not to quit on the
first failure in resolving ipv4only.arpa.