[inet] Set RTF_HOST only when removing a host route. JB#60276
INADDR_ANY, i.e., default routes do not get removed with the RTF_HOST
flag set. The flag must be set only for host routes. Otherwise the
INADDR_ANY route is not removed and in such case the removal of the
route will return a ok reply indicating that the route was removed.