timezone: Use zone.tab instead of zone1970.tab for completeness
[timezone] Use zone.tab instead of zone1970.tab for completeness. JB#55556

The zone1970.tab does not contain all of the regions and maps the
different ISO3166 codes under one region. This makes it impossible to
use the zone1970.tab for setting the WLAN region domain as in many cases
the region is not found in zone1970.tab and the regulatory domain is set
to default.

Instead, use the zone.tab originating also from IANA.org which has all
of the regions, each on separate line, and the ISO3166 code as well.
This makes it possible to always set the correct regulatory domain on
the WLAN, for instance.