network: keep failure state after disconnection
set_disconnected() function shall not prevent the service to process errors
by replacing failure state by idle.

For example, when a connection fails with a wrong passphrase, the user still
needs to be informed, via the agent. However, in this case, the associated
service is automatically disconnected, canceling the pending agent requests.
The invalid key error shall be reported, instead of a cancel error.

Signed-off-by: Emmanuel Vautrin <Emmanuel.VAUTRIN@cpexterne.org>