[clat] Stop prefix query when VPN runs over CLAT as default. JB#60276
When VPN is set to be the default service and it runs over CLAT (VPN
mode on) the periodic prefix query will fail because the DNS servers of
the VPN are only used. There is no DNS serving the request for CLAT
prefix, nor there is a route to the DNS server and, therefore, it is
better to simply drop the prefix query during the VPN is connected.

Enabling a DNS server of a transport would result in leaky DNS when VPN
is on because it would return, for example, an AAAA (IPv6) record for a
request and when VPN is IPv4 only the traffic would also bypass the VPN
as there is an active IPv6 connection, and, thus, the traffic would also
bypass CLAT and use the mobile data directly. That is reason enough for
using the approach in this commit because of simplicity and non-breaking
+ non-leaking.