[clat] Extend VPN mode: drop def IPv6 gw, add CLAT prefix IPv6 route. JB#60276
This change extends the VPN mode by dropping the default IPv6 gateway and
adding a CLAT prefix network route via the IPv6 gateway when VPN goes
over CLAT. This is to route only the CLAT prefixed traffic over IPv6 in
VPN mode.

The default IPv4 route for CLAT is managed by ConnMan core as the
ipconfig info is injected to the tracked service. Therefore, the adding
and removing of the IPv4 default route is disabled by default and can be
managed by adding a boolean to /etc/connman/clat.conf in group CLAT
(in square brackets, omitted here because of tag usage)

ClatDeviceSetDefaultRoute=true|false