[service] Ensure DHCPs are stopped when changing user. JB#60712
During user change the WIFI services of the previous user are unloaded.
If the change happens quite rapidly when both old and new user have the
same WIFI network saved or when the old user's WIFI network is still
waiting for DHCP reply it may be that the DHCPs are not getting stopped.
This might relate to the issue of wpa_supplicant getting a deauth
message from the AP as a result of multiple DHCP clients running on the
same device, i.e., there is already a pending connection attempt from
the same device.

This change ensures that both of the DHCPs are stopped. Both unref the
identifier (ipconfig / network) if the identifier was found from their
respective internal data structures.