[wifi] Change state also when interface is not valid. JB#60712
When enabling WiFi interface make sure that the checks on interface
validity also changes the state to undefined when error happens. This
may be one of the causes that WiFi is seemingly on but the status is not
updated to device or technology and, therefore, scan may be attempted
but ENOLINK is received as a result. This change should ensure that no
state change is missed in case of an error.