[connection] Fix remaining dual index support errors. Fixes JB#60943
Fix the errors caused by adding dual index support:
 - copy also the index of the other IP-family gw when replacing
 - enable searching the specific IP-family gateway
 - kernel netlink listens only IPv4 -> get IPv4 supporting gateway

Added also more debug to certain places to ease up following what is
going on.

The errors caused VPN cleanups not to work and in some cases, especially
with IPv6-disabling option set, VPN did not get set as default route
and/or the old default route was not sent. This resulted in
non-functioning networking, including disabled DNS when VPN was
connected for the 2nd and ongoing times.