Be consistent about debug logging service and IP config type.
In a number of DBG log instances, there are inconsistencies in what and
how a service and its IP configuration type are logged. In some places,
it is just pointers and enumeration values. In other places, it is
pointers, identifiers, and strings without enumerated values.

This makes several such instances consistent, logging the service
pointer, the service instance name, the IP configuration type enumerated
value and its string equivalent.