Terminate all online check handling in one function.
This modifies 'portal_manage_status' by moving the unconditional
transition of the service to the 'online' state for which a HTTP-
based reachability check was successfully completed to the callback
provided to '__connman_wispr_start', specifically,
'complete_online_check'. This places all handling, both failure and
success, of HTTP-based reachability checks in a single, auditable
location in a single function in a single source file.