gweb: Add TCP connect timeout member and getter/setter.
This adds a TCP connect timeout member, getter, and setter. Collectively,
these manage the TCP connection timeout, in milliseconds, for a specified
GWeb object. Connections that take longer than this will be aborted. A
value of zero ('0') indicates that no explicit connection timeout will be
used, leaving the timeout to the underlying operating system and its
network stack.