main: Introduce the 'OnlineCheckConnectTimeout' setting.
This introduces the 'OnlineCheckConnectTimeout' setting which is the
time, in decimal seconds (for example, 12.3), to wait for a successful
TCP connection to the host associated with "OnlineCheckIPv4URL" or
"OnlineCheckIPv6URL". Connections that take longer than this will be
aborted. The default value is zero ('0') which indicates that no explicit
connection timeout will be used, leaving the timeout to the underlying
operating system and network stack.