service: Group online check state by IP configuration type.
As the implementation of the continuous online check mode (also known
as, EnableOnlineToReadyTransition) evolves, there is an increasing
need for spatial locality of online check state by IP configuration
type.

Consequently, rather than having a collection of individual data
members in the service structure decorated with '_ipv4' and '_ipv6'
and accessing them one-by-one, group them into a common
structure. Then, instantiate that structure twice in the service
structure, once for IPv4 and once for IPv6. This allows for the
desired spatial locality.